# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CountMax Common API",
        "contact": {
            "name": "CountMax",
            "url": "https://git.countmax.ru/countmax/commonapi",
            "email": "1020@watcom.ru"
        },
        "license": {},
        "version": "2.0"
    },
    "paths": {
...
        "/v2/videochecks/configs": {
            "get": {
                "description": "get slice of customer videocheck configuration with offset, limit parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cm_info"
                ],
                "summary": "Get all videocheck configurations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "default=0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default=10",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/infra.VideoCheckResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "inserts customer videocheck configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cm_info"
                ],
                "summary": "Inserts new customer's videocheck configuration",
                "parameters": [
                    {
                        "description": "New videocheck configuration",
                        "name": "videocheckConfig",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.VideocheckConfig"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/infra.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    }
                }
            }
        },
        "/v2/videochecks/configs/{pid}": {
            "get": {
                "description": "get customer videocheck configuration by customerID parameter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cm_info"
                ],
                "summary": "Get customer's videocheck configuration",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Code 1S",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.CustomerConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "updates customer videocheck configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cm_info"
                ],
                "summary": "Update customer's videocheck configuration",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Code 1S",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New videocheck configuration",
                        "name": "videocheckConfig",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.VideocheckConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/infra.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "delete customer videocheck configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cm_info"
                ],
                "summary": "Delete customer's videocheck configuration",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Code 1S",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/infra.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.CustomerConfig": {
            "type": "object",
            "properties": {
                "customerId": {
                    "type": "integer"
                },
                "customerName": {
                    "type": "string"
                },
                "customerTypeId": {
                    "type": "integer"
                },
                "sdCreatorId": {
                    "type": "integer"
                },
                "sdDestination": {
                    "type": "string"
                },
                "sdServiceId": {
                    "type": "integer"
                }
            }
        },
...
        "domain.VideocheckConfig": {
            "type": "object",
            "properties": {
                "localCam": {
                    "type": "boolean"
                },
                "localFtp": {
                    "type": "boolean"
                },
                "localServer": {
                    "type": "boolean"
                },
                "options": {
                    "type": "string"
                },
                "projectId": {
                    "type": "integer"
                }
            }
        },
        "domain.VideocheckConfigs": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/domain.VideocheckConfig"
            }
        },
...
        "infra.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "application-specific error code",
                    "type": "integer"
                },
                "error": {
                    "description": "application-level error message, for debugging",
                    "type": "string"
                },
                "status": {
                    "description": "user-level status message",
                    "type": "string"
                }
            }
        },
        "infra.HTTPError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "object"
                }
            }
        },
...
        "infra.ResultSet": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "infra.SuccessResponse": {
            "type": "object",
            "properties": {
                "httpcode": {
                    "description": "http response status code",
                    "type": "integer"
                },
                "message": {
                    "description": "text of message",
                    "type": "string"
                },
                "status": {
                    "description": "user-level status message",
                    "type": "string"
                }
            }
        },
        "infra.VideoCheckResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/domain.VideocheckConfigs"
                },
                "result_set": {
                    "type": "object",
                    "$ref": "#/definitions/infra.ResultSet"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}